+++ client/src/app/components/rdio-scanner/admin/admin.service.ts
@@ -295,7 +295,6 @@ export interface Options {
         googleAPIKey?: string;
         googleCredentials?: string;
         assemblyAIKey?: string;
-        assemblyAISpeechModel?: string;
         assemblyAIWordBoost?: string[];
         hallucinationPatterns?: string[];
         hallucinationDetectionMode?: string;
@@ -1227,7 +1226,6 @@ export class RdioScannerAdminService implements OnDestroy {
                 googleAPIKey: this.ngFormBuilder.control(transcriptionConfig?.googleAPIKey || ''),
                 googleCredentials: this.ngFormBuilder.control(transcriptionConfig?.googleCredentials || ''),
                 assemblyAIKey: this.ngFormBuilder.control(transcriptionConfig?.assemblyAIKey || ''),
-                assemblyAISpeechModel: this.ngFormBuilder.control(transcriptionConfig?.assemblyAISpeechModel || ''),
                 assemblyAIWordBoost: this.ngFormBuilder.control(
                     (transcriptionConfig?.assemblyAIWordBoost || []).join('\n')
                 ),